Add initial implementation

+++ src/map.h
@@ -14,7 +14,7 @@ class Map {
 public:
 
     struct single_landmark_s {
-        int id_i; // Landmark ID
+        size_t id_i; // Landmark ID
         float x_f; // Landmark x-position in the map (global coordinates)
         float y_f; // Landmark y-position in the map (global coordinates)
     };
